Add tool invocation logging support
Updated the command-line tool framework to add support for logging information
about the launch and completion of command-line tools to a specified file or
set of files.  Launch messages will include the command name and all of the
arguments provided on the command line, as well as the values of any arguments
obtained from a properties file, the time the command was invoked, and the
name of the account on the underlying system that launched the tool.
Completion messages will include the exit code and/or exit message, if
available (although they may not be available if the tool was interrupted
before it completed).

This functionality is primarily intended for use with Ping Identity server
products, and by default, logging will only be performed for those tools (and
also by default, only for a subset of tools that can alter server
configuration, state, or data).  However, it is possible to enable invocation
logging for any tool written with the LDAP SDK command-line tool API by
setting an environment variable or JVM property that the command-line tool
framework can use to locate a logging properties file, and by defining an
appropriate configuration in that properties file.