Temporarily disable tool invocation logging
The command-line tool framework includes support for tool invocation logging,
in which the tool can record log messages whenever a tool is launched, and
when that tool completes.  This is primarily intended for use in tools that
ship with Ping Identity server products, and the corresponding server-side
support is not yet implemented.

To avoid partially exposing this feature, the LDAP SDK support for this
functionality has been temporarily disabled by commenting out the code that
looks for an environment variable with the server instance root.  The code
that looks for a Java system property is still there, so the unit tests are
still run, and the invocation logging can still be used if you really want it,
but it will no longer happen automatically for certain LDAP SDK tools until
we're actually ready to expose this capability.