Re-enable tool invocation logging
The command-line tool framework has the ability to log messages
whenever certain command-line tools are invoked and when they complete.
This feature is primarily intended for use with Ping Identity server
products, but the server-side support for this feature has not yet been
implemented.

To avoid partially exposing this feature in server products that rely
on the recently released 4.0.1 version of the LDAP SDK, the code that
can automatically enable tool invocation logging based on the presence
of a specific environment variable was temporarily commented out for
the 4.0.1 release.  Now that the 4.0.1 version of the LDAP SDK has been
released, that code has been reinstated.