Allow hiding argument IDs and subcommand names
When creating an argument, you can only specify at most one long
identifier string and one short identifier character, and you have to
specify at least one of them.  However, you can also specify additional
long and short identifiers that can also be used for that argument on
the command line.  This is useful because it can make it easier for
users to request an argument without having to remember exactly what
the right identifier is.  For example, if a tool creates an LDIF file,
a user might not remember off the top of their head if they need to
specify the argument as '--ldifFile' or '--ldif-file' or '--filename'
or '--outputFile' or '--output-file'.  If the tool defines multiple long
identifiers for that argument, then the user could provide any of those
identifiers to get the desired result.  This helps make the tool more
user-friendly and convenient to use.

In the past, all of the long and short identifiers for an argument were
always included in the usage information.  This can be useful because it
ensures that the user knows all of the ways that argument can be
requested, but it can also cause the output to appear cluttered,
especially for arguments with a lot of identifiers.  To address this,
it is now possible to indicate that additional long and short
identifiers should be hidden.  This means that the tool will still
accept the provided identifier as an alternate way to request the
argument, but it won't appear in usage information.

Similarly, for tools that use subcommands, you can provide additional
names for each subcommand that can also be used to request that
subcommand.  When adding an additional subcommand, you can now indicate
whether that additional subcommand name should be hidden so that it is
still allowed on the command line but will be excluded from usage
information.

All of the command-line tools provided with the LDAP SDK have been
updated so that all additional long and short identifiers, and all
additional subcommand names, will be marked as hidden.  All of those
additional argument identifiers and subcommand names are still available
for use, but they will be excluded from usage information so that the
output is less cluttered.