Fixed an SSLSocketVerifier connection leak
If a connection is configured with an SSLSocketVerifier and that
verifier rejects the connection for some reason, the LDAP SDK
previously left that socket established and could keep it open until
the JVM shut down or until it was closed by the server.  The LDAP SDK
will now properly close the connection.