Improve LDAP debug messages
Updated the LDAP SDK debug logging to improve the quality of debug
messages generated when sending an LDAP request.  The debug message
will now include the internal connection ID, connection name,
connection pool name, and address and port of the server to which the
connection is established, and the LDAP message ID for the message that
is being sent.  The debug messages generated for abandon and unbind
requests are now more consistent with those generated for other types
of requests, and we do a better job of creating debug messages for SASL
bind requests.

In addition, we have also improved the quality of the message in the
exception that will be thrown if a connection operating in
non-synchronous mode attempts to register multiple response acceptors
for the same message ID on the same connection, which suggests an
attempt to send multiple concurrent requests with the same message ID.