Correct the UTC time universal BER type
Fixed a bug in ASN1Constants that used an incorrect BER type for UTC
time elements.  The correct value is 0x17 (one less than the type for
the generalized time type), but an incorrect value of 0x19 (one more
than the generalized time type) was used in the initial commit.