Retain encoded certificates and private keys
Updated the X509Certificate and PCKS8PrivateKey classes to retain the
byte arrays containing the encoded representation of the certificate or
private key.  The byte array is made available to callers, and the
X509Certificate class has been updated to provide methods for getting
the certificate's SHA-1 or 256-bit SHA-2 fingerprint.