Fix certificate and private key bugs
Fixed a couple of bugs in the code for building and parsing X.509
certificates and PKCS#8 private keys.  The certificate code did not
properly handle the case in which a certificate might not include a
version (which should be omitted v1 certificates), and neither the
certificate code nor the public key code properly handled the
possibility of nonexistent algorithm parameters.