Fixed a bug in sending a simple bind request
When the LDAP SDK is configured to use synchronous mode, if the call to
send a bind request message fails, the LDAP SDK could silently consume
the exception rather than propagating it back to the caller, and would
then proceed to wait for a response that won't come because the request
probably didn't get sent.  This could cause the thread to be blocked
for the entire response timeout period before an exception is thrown
back to the caller, and when that happens, the exception about a
response timeout could be misleading because the request may not have
actually been sent.

Most of the time, there should be a very low risk of getting an
exception when trying to send a message because that message will go
into the socket's transmit buffer, and it is highly unlikely that the
message is too big to fit in the transmit buffer (and because the
connection is operating in synchronous mode, it is also highly unlikely
that the transmit buffer has anything else in it).  This means that in
most cases, the call to send the bind request message will succeed
almost instantaneously and the LDAP SDK will block while waiting on
the response.  If a problem the occurs that causes the request to not
actually be sent, then the call to read the response will fail with the
details of the send failure.

However, the problem could arise if the bind request is the first
operation on a newly-established secure connection backed by an
SSLSocket in which the TLS negotiation has not yet completed.  Unlike a
regular socket, an SSLSocket will block while trying to send data over
a connection that has not yet completed TLS negotiation.  If the TLS
negotiation fails or times out, then the application wouldn't have
received any indication of a problem until after the full response
timeout period has elapsed after the problem was first detected, and
then the exception that eventually got thrown would erroneously suggest
that the problem was in not getting the expected response within an
acceptable period of time rather than failing to send the request in
the first place.