Added an LDAPConnection.closeWithoutUnbind method
Updated the LDAP SDK to add a new LDAPConnectionc.closeWithoutUnbind
method that can be used to close a connection without first sending an
unbind request.  Although this is generally discouraged, there may be
cases in which this is useful (for example, if there is reason to
believe that attempting to send data on the connection will fail or
block).  Connection pool implementations have been updated so that
when closing a connection that is believed to be invalid, the
connection will be closed without an unbind request.