Improved OID validation support.
The LDAP SDK already had support for parsing strings as OIDs, and for
determining whether a string represented a valid OID.  However, this
validation was not strictly compliant with the requirements of numeric
OIDs.  In particular, a strictly valid numeric OID should also meet the
following requirements that had not previously been checked:

* It must have at least two components.
* The first component must have a value of 0, 1, or 2.
* If the first component has a value of 0 or 1, then the second
  component must not be larger than 39.

To retain backward compatibility but also provide support for strict
validation, the OID class has been updated to keep the
isValidNumericOID method as-is, but add a new isStrictlyValidNumericOID
method that performs all of the additional strict checking.

In addition, a new OIDArgumentValueValidator class has been added,
which can be added to an Argument (presumably a StringArgument) to
ensure that the argument only accept values that are numeric OIDs.
This validator can be configured to operate in either strict or
non-strict mode.