Added PasswordReader.readPasswordChars
The PasswordReader utility class already provides a readPassword method
that can be used to read a password from the console and to get that
password as a byte array.  There is now a new readPasswordChars method
that can get the password as a character array as an alternative to a
byte array.