Added a byte-for-byte diff option
Updated the Entry.diff option to add an option to perform byte-for-byte
value comparisons.  This option makes it possible to detect changes
that result in a different value that is logically equivalent to the
previous value (for example, a value differing only in capitalization
in a case-insensitive attribute).  In addition, the LDAPPersister class
has been updated to make it possible to use byte-for-byte comparisons
when identifying changes to an object.