Update manage-certificates to use JKS by default
Updated the manage-certificates tool so that it will use JKS instead
of PKCS #12 as the default keystore type.  Although PKCS #12 is the
preferred format for security reasons, and although PKCS #12 will be
the JVM's default keystore type starting with Java 9, Java SE 7
does not support adding trusted certificate entries to a PKCS #12
keystore.  This capability is available as of Java SE 8, and
manage-certificates can be used to add trusted certificate entries
to PKCS #12 keystores when run on a Java 8 VM, but because the LDAP
SDK needs to support Java SE 7, it will use JKS as the default
keystore type that will be used if you don't explicitly specify a
type when creating a new keystore.