Added improved SO_TIMEOUT debugging
Updated the LDAP SDK to add improved debugging around the SO_TIMEOUT
socket option.  Whenever the SO_TIMEOUT is set on a connection, a
debug message is generated with a debug type of CONNECT and a level
of INFO.  Whenever an LDAP request is sent and a connection is
available, the SO_TIMEOUT value is included in the existing LDAP
request debug message.