Updated a manage-certificates example
Updated the manage-certificates tool to fix an example that
demonstrated the process for importing a certificate into a
keystore.  The example would have attempted to import a certificate
without a private key into a PKCS #12 keystore, which works in Java
8 but is not supported in Java 7.  The example has been updated to
use a JKS keystore, which works in both Java 7 and 8.