Fix a manage-certificates list-certificates bug
The list-certificates subcommand of the manage-certificates tool
will attempt to verify the signature for each certificate in the
keystore.  To do this, it needs to retrieve the issuer certificate,
and if it can't find that issuer certificate in the keystore, it
will report a warning.  Previously, the tool would only look for
the issuer certificate in the keystore's trusted certificate
entries, which means that it wouldn't find it if the issuer
certificate were only included in the keystore as part of the
certificate chain stored in a private key entry.  The tool will
now properly look for the issuer certificate in private key
entry certificate chains in addition to trusted certificate entries.

The check-certificate-usability subcommand also attempts to validate
the signatures of each certificate in the chain, but it was not
affected by this bug because it uses a different mechanism for
getting the issuer certificate.