Allow exporting a chain to separate files
Updated the manage-certificates tool so that the export-certificate
subcommand now includes a --separate-file-per-certificate argument
that can be used if both the --export-certificate-chain and the
--output-file arguments are also given.  If all of these arguments
are provided, and if there are multiple certificates to be exported,
then the file used to hold the first certificate will be the path
specified using the --output-file argument with ".1" appended to it,
the file for the second certificate will be the specified path with
".2" appended to it, and so on.  If the
--separate-file-per-certificate argument is not given, then all of
the certificates will be concatenated into the same file.