Make Attribute.hasValue more efficient
Updated the Attribute.hasValue method so that it will be more
efficient for attributes with multiple values, and especially for
attributes with a large number of values, and especially for
attributes that use more complicated matching rules, like attributes
with a distinguished name or generalized time syntax.

Previously, when attempting to determine whether an attribute had a
particular value, the LDAP SDK would simply iterate through all of
the values of that attribute and use the MatchingRule.valuesMatch
method for the associated matching rule.  In most cases, this would
require normalizing the assertion value for every comparison.  Now,
we use the new MatchingRule.matchesAnyValue method, which will try
to only normalize the assertion value once for the entire attribute
rather than repeating it for each value.

This should not yield any noticeable improvement (or degradation)
for attributes that have only a single value, but it should
improve performance for attributes with multiple values, and the
benefit will increase as the number of values increases.

This will also yield a similar improvement for the
Filter.matchesEntry method for equality filters that target
attributes with multiple values.