Allow running a subset of unit tests
Updated the LDAP SDK build process to make it possible to run a
selected subset of unit tests.  To accomplish this, two new system
properties have been defined:

* The "test.classes" property allows you to specify the fully
  qualified names of one or more unit test classes to be invoked.
  If multiple class names are specified, then they should be
  separated by commas.

* The "test.packages" property allows you to specify the fully
  qualified names of one or more unit test packages containing the
  classes to be invoked.  If any test packages are specified, then
  all tests in those packages, and in all of their subordinate
  packages, will be invoked.  If multiple packages are specified,
  then they should be separated by commas.

For example, running:

    ./build.sh test -Dtest.classes=com.unboundid.util.Base64TestCase

will cause just the tests contained in the Base64TestCase class to
be invoked.

If neither of these properties is specified, then the full set of
unit tests will be invoked.