Improve the prompt trust manager
Updated the LDAP SDK's prompt trust manager (which can interactively
prompt the user about whether to trust a certificate) to provide
better output formatting, more warnings about potential reasons that
a certificate chain should not be trusted (for example, because it
includes a certificate with an invalid signature, because the chain
is incomplete, because there is a problem with one of the
extensions, etc.), and whether a server certificate contains the
address that the client used to connect to that server, either in
the CN attribute of the certificate subject or in a subject
alternative name extension.  The LDAP command-line tool framework
has been updated so that tools written with that framework can use
the new address-aware validation.