Better handle stalled TLS negotiation
Updated the way that the LDAP SDK sends messages so that it is more
resilient to stalls in the TLS negotiation process.  Java's
SSLSocket implementation has the potential to block indefinitely
until TLS negotiation has completed, and the only way to break out
of that stall is to use the SO_TIMEOUT socket option.

This was primarily an issue for connections operating in
non-synchronous mode, as connections in synchronous mode already
used the SO_TIMEOUT socket option in order to enforce response
timeouts.