Avoid blocking multiple threads on pool close
When closing an LDAPConnectionPool or LDAPThreadLocalConnectionPool,
the health check thread will be shut down, and the call to do that
will block until the health check thread has actually stopped.

Once the health check thread has been stopped, the connection pool
will close any connections that are established and idle.  However,
this will not affect any connections that are currently checked
out.  Instead, those connections will be closed whenever they are
released back to the pool.  This calls the same close method that
external code can use to shut down a pool, which mean that if the
health check thread has not yet finished running, the thread
releasing the connection back to the pool will also be blocked.

It doesn't make sense to have multiple threads blocked on the
health check thread shutdown, so the connection pools have been
updated so that only the first attempt to close the pool has the
potential to block on the health check thread shutdown.  Any
subsequent calls to the close method (including internal calls made
when releasing a connection back to the pool that has already
started the shutdown process) will not have to wait for the health
check thread to stop running.