Updated the get entry health check
Updated the get entry LDAP connection pool health check to add
support for invoking the health check after a pooled connection has
been authenticated.  This includes immediately after a newly-created
connection has been authenticated (if pooled connections should be
authenticated), after a call to the connection pool's
bindAndRevertAuthentication method, and after a call to the
connection pool's releaseAndReAuthenticateConnection method.  It
also clarifies that if the health check is invoked for a
newly-created connection, that check will be performed before the
connection is authenticated.

The LDAPConnectionDetailsJSONSpecification class has also been
updated to make it possible to configure this behavior for
connection pools created from information contained in a JSON
object.