Fixed the get entry health check after exception
Fixed a bug in the GetEntryLDAPConnectionPoolHealthCheck that would
cause it to behave incorrectly when checking the validity of a
connection after an exception.  The intended behavior was to only
search for the target entry if the exception result code was one
that indicated that the connection may no longer be valid, and to
only fail the health check of that target entry could not be
retrieved.  But instead, it exhibited the incorrect behavior of
always failing the health check without performing a search if the
exception result code indicated that the connection may not be
usable anymore, and to always perform a search to determine the
connection's validity only if the exception was one that did not
suggest that the connection may be invalid.