Add a FixedRateBarrier.await(int) method
Updated the FixedRateBarrier class to add an await method that takes
an integer argument that represents the number of occurrences for
the associated action to include in the await operation.  The
existing await() method that doesn't take any arguments is now
functionally equivalent to calling await(1).

This new version of the await method can be useful in cases in which
you want to perform rate limiting for something in which it is not
feasible or desirable to call the await() method for each
occurrence.  For example, if you want to impose rate limiting for
data being written to an output stream, it doesn't make sense to
call await() for each byte that is written.  Instead, you can now
call await(int) with a value that is the number of bytes you want
to write, and the method will not return until it is appropriate
to write the requested number of bytes.  Note that the value
provided to the await(int) method must not exceed the perInterval
value currently in use by the rate limiter, and it should ideally be
a relatively small percentage of the perInterval value to ensure
that the barrier can achieve a rate that is consistently close to
the target.