Fix a failing unit test
The MonitorManagerTestCase unit test, which verifies that the LDAP
SDK can interact with UnboundID/Ping Identity Directory Server index
monitor entries, started failing against builds of a
not-yet-released version of the Directory Server.  The test
assumed that each index monitor entry would include a value for the
ds-index-backend-base-dn attribute, but that is no longer the case.