Updated a manage-certificates argument
Updated the manage-certificates tool to change the primary name for
the --replace-existing-certificate argument in the
generate-certificate-signing-request subcommand to be
--use-existing-key-pair, and updated the tool usage to make it
clearer that using that argument will not cause the keystore to be
updated.  Because arguments support multiple long identifiers, the
former --replace-existing-certificate argument is still available
and will have exactly the same effect, so backward compatibility is
not affected in any way.

Also updated the manage-certificates tool to be more consistent when
using the phrase "key pair".  It was used as a single word
("keypair") in some instances, and as two words in others.  All uses
have been updated to use the two-word variant, which appears to be
the predominant form.