Another passphrase-encrypted stream header update
Updated the passphrase-encrypted stream header to make it possible
to read the header from an input stream or decode it from a byte
array without providing a passphrase.  This makes it possible to
obtain the contents of the header without having access to the
passphrase, which can be useful if information inside the header
(for example, the key identifier) may be needed to determine which
passphrase to use.  If the header was read or decoded without a
passphrase, then the MAC will not be validated, and it will not be
possible to use that header to perform encryption or decryption.