Fix a bug in manage-certificates validity
Fixed a bug in the manage-certificates tool in which it could use
incorrect values for the notBefore and notAfter elements in a
self-signed certificate when replacing an existing certificate in
the key store.  In such cases, manage-certificates would always use
the current time as the start time and a duration of one year
instead of the values of the --validity-start-time and --days-valid
arguments.

Also, updated the generate-self-signed-certificate so that it is
possible to use --use-existing-key-pair as an alternative name for
the --replace-existing-certificate argument.  This provides better
consistency with the generate-certificate-signing-request
subcommand.