Better summarize-access-log encryption support
Updated the summarize-access-log tool to improve support for
examining log files that the server currently has open.  An
encrypted log file that the server still has open is unlikely to be
fully decryptable because the file won't end with a valid
properly-padded block.  This commit adds a message that explains
the error, and then proceeds with an attempt to summarize the log
content that it was able to read successfully.