Add forName methods to enum classes
Update a number of enum classes to add forName methods that make it
possible to retrieve a value of the enum using its name.  These
forName methods are more forgiving than the standard valueOf method
that takes a string, since the standard value will throw an
IllegalArgumentException when provided with a string that does not
exactly match the value returned by the enum's name() method.  The
forName methods are case-insensitive, can handle various forms of
the name (with words separated by underscores, dashes, or no
delimiter), and return null rather than throwing an exception if
the requested name isn't defined.