Fix basic constraints path length interpretation
Fixed a bug in how manage-certificates handles the path length
element of a basic constraints extension.  It was originally written
with the incorrect assumption that the value specified the maximum
number of certificates allowed in the entire certificate chain.
However, the correct interpretation is that it specifies the maximum
number of intermediate certificates that are allowed to exist
between that issuer certificate and the subject certificate at the
head of the chain.  The manage-certificates tool has been updated so
that arguments that allow you to specify a basic constraints path
length now allow a value of zero (which is explicitly documented as
valid), and to correct the validation performed in the
check-certificate-usability subcommand.