Optionally allow SHA-1 signatures for issuer certs
Updated the manage-certificates check-certificate-usability
subcommand to add an --allow-sha-1-signature-for-issuer-certificates
argument that will prevent the tool from exiting with an error just
because one of the issuer certificates has a SHA-1-based signature.
SHA-1 is considered weak, and many TLS libraries no longer support
entity certificates with SHA-1-based signatures, but there are still
some commercial authorities that have for whatever reason not
started using new issuer certificates with a SHA-2-based signature
algorithm.  When using the new argument, these SHA-1-based
signatures will still be called out, but the tool won't exit with an
error code just because of them.