Add a task to reload HTTP connection handler certs
Added client-side support for a task that may be used to invoke a
Ping-proprietary administrative task for requesting that the server
reload the key and trust manager providers associated with any HTTP
connection handlers that provide support for HTTPS.  This can cause
any changes to the certificates in those stores to be applied to
the running instance without the need to disable and re-enable the
connection handler or to restart the server.  Note that this reload
does have the potential to interfere with the ability for clients to
resume a TLS session created before the reload was invoked.