Specify character sets by object instead of name
Updated the LDAP SDK so that when converting between strings and
byte arrays, the character set to use is provided as a
java.nio.charset.Charset object instead of a String that contains
the name of the desired character set.  This may improve
performance, especially in cases where a lot of conversion happens
all at once, and it can also help improve reliability.

Also updated a handful of cases in which the LDAP SDK converted
between a string and a byte array without explicitly specifying a
character set.  There were only a limited number of such
occurrences, and generally only when it is known that the string
is comprised entirely of ASCII characters, but the new approach
should still provide better performance and reliability.