Improve Error handling in the In-Memory DS
Updated the in-memory directory server to better handle any
java.lang.Error (and subclasses) instances that may be thrown during
processing.  Previously, if an Error was thrown, the thread handling
the client connection could exit without returning an error response
or closing the connection, which could cause a client to wait
indefinitely for a response that isn't going to come.  It will now
return an error (if appropriate for the operation type) and close the
connection.