Add support for per-request referral connectors
Added an LDAPReqest.setReferralConnector method that makes it
possible to specify a referral connector that should be used for
that request, overriding a referral handler inherited from the
LDAPConnection or LDAPConnectionOptions that would have otherwise
been used.

Also added a RetainConnectExceptionReferralConnector class that makes
it easier to obtain an exception that was caught while trying to
establish a connection for the purpose of following a referral.