Fix overlooked bind state updates
When processing a bind request, the LDAP SDK updates some client-side
state information for the connection.  In particular, it retains
information about the bind request that may be needed when
re-establishing an existing connection or creating a new connection
from an existing one (for example, for the purpose of following a
referral or populating a connection pool).  It may also be configured
to maintain a cached copy of the server's schema, and that cached
copy should get updated after a successful bind because different
users may theoretically have different views of the schema.

A bug was identified in which this client-side connection state
information may not get updated properly when processing a bind
request through either the LDAPConnection.processOperatino method or
the AbstractConnectionPool.processRequests method.  This could cause
the LDAP SDK to incorrectly believe that a connection was not
authenticated when it actually was.  This has been corrected.