More website-related updates
Updated a couple more references to LDAP.com so that they now point
to the GitHub project as the main LDAP SDK home page.  Also updated
the generated documentation so that the documentation that will be
published on LDAP.com now has the same basic appearance as the
offline content that is included in the LDAP SDK download.