Add a streamfile value pattern component
Updated the LDAP SDK to add a streamfile value pattern component,
which behaves like the sequentialfile component except that it does
not try to load the entire contents of the file into memory.
Instead, it uses a background thread to read data into a fixed-size
queue, exiting the background thread and closing the file if the
queue becomes full and an attempt to add new data to it blocks for
some period of time.  This makes it suitable for working with very
large files without needing to worry too much about the performance
impact of reading from the file on demand.