Allow value patterns in modrate; add new patterns
Updated the modrate tool to add a new valuePattern argument that
makes it possible to provide a value pattern when generating the
values for the modifications.  This is an alternative to the existing
valueLength and characterSet arguments and allows for more
flexibility.

Also, added a few new supported value pattern component types,
including:

* A "random" component, which can be used to generate a string of
  a specified number of characters from a given character set.

* A "timestamp" component, which can be used to generate timestamp
  values, representing either the current time or a randomly selected
  time within a given range, in a variety of formats.

* A "uuid" component, which can be used to generate UUID (universally
  unique identifier) values.