Guard against clear PWs in tool invocation log
Updated the tool invocation logger to better guard against having
passwords show up in the clear in the log.  Although any argument
that is marked sensitive should have its value redacted, if a
sensitive argument is accidentally not marked as sensitive then its
value will show up in the log in the clear.  To help prevent this
from happening with passwords, the tool invocation logger will now
look for any argument whose name suggests that the value is a
password, and will redact its value if it does not already appear to
be redacted.