Fix a connection pool double-bind bug
Fixed a problem in the LDAPConnectionPool and
LDAPThreadLocalConnectionPool classes that was triggered by creating
a pool from an already-established connection instead of creating it
from a server set and bind request.  Internally, the LDAP SDK would
create a server set and bind request from the provided connection,
but it created the server set in a way that would cause it to
authenticate the connection as part of establishing it.

Under most circumstances, this isn't a huge problem, although it
does cause the pool to send a second, duplicate bind request to the
server when creating a new connection for use in the pool.  But it
does cause a more significant problem when calling the pool's
setBindRequest method because although the pool's bind request is
successfully updated, the server set was still creating connections
that were authenticated with the old credentials before trying to
bind with the new bind request.  If those old credentials were no
longer valid, then the initial bind attempt would fail, which would
cause the entire connection attempt to fail.

The pool implementations have now been updated so that when creating
a pool from an already-established connection, the server set that
gets created won't try to authenticate the connections.  This will
improve performance when establishing a new connection because only
one bind request will be processed, and it will make setBindRequest
work as expected.