Add additional multi-byte character unit tests
Added additional unit test coverage for filters, DNs and RDNs, and
matching rules for characters comprised of multiple bytes.  In this
case, the primary addition is for a country flag emoji, which is
actually comprised of multiple characters (for example, the U.S. flag
emoji is comprised of the "Regional Indicator Symbol Letter U"
character followed by the "Regional Indicator Symbol Letter S"
character.